Update memnet.py
Hello,

I made a small modification to the calculation of the locationed memory in this model. The previous implementation was not using the absolute distance context word - aspect as specified in the paper.

Hope you find this useful. If so, it will also require to include this change in the input_colse of the train.py file
'memnet': ['text_raw_without_aspect_indices', 'aspect_indices', 'text_left_with_aspect_indices']